split tokens/consts from implementation
