+++ src/packages/user/user/collection/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_USER_COLLECTION_ALIAS = 'Umb.Collection.User';
