+++ src/packages/user/user/collection/repository/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_USER_COLLECTION_REPOSITORY_ALIAS = 'Umb.Repository.UserCollection';
