+++ src/packages/user/user/collection/repository/manifests.ts
@@ -1,6 +1,5 @@
+import { UMB_USER_COLLECTION_REPOSITORY_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_USER_COLLECTION_REPOSITORY_ALIAS = 'Umb.Repository.UserCollection';
 
 const repository: ManifestRepository = {
 	type: 'repository',
