+++ src/packages/user/user/collection/views/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_COLLECTION_VIEW_USER_TABLE = 'Umb.CollectionView.User.Table';
+export const UMB_COLLECTION_VIEW_USER_GRID = 'Umb.CollectionView.User.Grid';
