+++ src/packages/user/user/collection/views/index.ts
@@ -1,2 +1,1 @@
-export { UMB_COLLECTION_VIEW_USER_TABLE } from './manifests.js';
-export { UMB_COLLECTION_VIEW_USER_GRID } from './manifests.js';
+export * from './constants.js';
