+++ src/packages/user/user/invite/repository/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_INVITE_USER_REPOSITORY_ALIAS = 'Umb.Repository.User.Invite';
