+++ src/packages/user/user/invite/repository/manifests.ts
@@ -1,6 +1,6 @@
+import { UMB_INVITE_USER_REPOSITORY_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
 
-export const UMB_INVITE_USER_REPOSITORY_ALIAS = 'Umb.Repository.User.Invite';
 const inviteRepository: ManifestRepository = {
 	type: 'repository',
 	alias: UMB_INVITE_USER_REPOSITORY_ALIAS,
