+++ src/packages/user/user/repository/avatar/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_USER_AVATAR_REPOSITORY_ALIAS = 'Umb.Repository.User.Avatar';
