+++ src/packages/user/user/repository/avatar/index.ts
@@ -1,2 +1,2 @@
 export { UmbUserAvatarRepository } from './user-avatar.repository.js';
-export { UMB_USER_AVATAR_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_USER_AVATAR_REPOSITORY_ALIAS } from './constants.js';
