+++ src/packages/user/user/repository/change-password/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_CHANGE_USER_PASSWORD_REPOSITORY_ALIAS = 'Umb.Repository.User.ChangePassword';
