+++ src/packages/user/user/repository/change-password/index.ts
@@ -1,2 +1,2 @@
 export { UmbChangeUserPasswordRepository } from './change-user-password.repository.js';
-export { UMB_CHANGE_USER_PASSWORD_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_CHANGE_USER_PASSWORD_REPOSITORY_ALIAS } from './constants.js';
