+++ src/packages/user/user/repository/change-password/manifests.ts
@@ -1,6 +1,5 @@
+import { UMB_CHANGE_USER_PASSWORD_REPOSITORY_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_CHANGE_USER_PASSWORD_REPOSITORY_ALIAS = 'Umb.Repository.User.ChangePassword';
 
 const changePasswordRepository: ManifestRepository = {
 	type: 'repository',
